# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Aplikasi Booking Kost API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/kamar": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kamar-controller"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kamar-controller"
                ],
                "parameters": [
                    {
                        "description": "Data Kost",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.KamarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.KamarResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/kamar/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kamar-controller"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Kamar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Kamar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateKamarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.KamarResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kamar-controller"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Kamar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.APIResponse"
                        }
                    }
                }
            }
        },
        "/kost": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kost-controller"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ViewKostResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kost-controller"
                ],
                "parameters": [
                    {
                        "description": "Data Kost",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddKostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.KostResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/kost/kamar": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kost-controller"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.KamarKostReponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/kost/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kost-controller"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Kost",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Kost",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateKostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.KostResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kost-controller"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Kost",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "request.AddKostRequest": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "namaKost": {
                    "type": "string"
                },
                "pemilik": {
                    "type": "integer"
                },
                "typeKost": {
                    "type": "string"
                }
            }
        },
...
        "request.KamarRequest": {
            "type": "object",
            "properties": {
                "hargaKamar": {
                    "type": "integer"
                },
                "kostId": {
                    "type": "integer"
                },
                "namaKamar": {
                    "type": "string"
                },
                "statusKamar": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateKamarRequest": {
            "type": "object",
            "properties": {
                "hargaKamar": {
                    "type": "integer"
                },
                "kostId": {
                    "type": "integer"
                },
                "statusKamar": {
                    "type": "string"
                }
            }
        },
        "request.UpdateKostRequest": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "namaKost": {
                    "type": "string"
                },
                "pemilik": {
                    "type": "integer"
                },
                "typeKost": {
                    "type": "string"
                }
            }
        },
...
        "response.GetKamarResponse": {
            "type": "object",
            "properties": {
                "hargaKamar": {
                    "type": "integer"
                },
                "nomorKamar": {
                    "type": "string"
                },
                "statusKamar": {
                    "type": "string"
                }
            }
        },
        "response.KamarKostReponse": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "detailKamar": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.GetKamarResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "namaKost": {
                    "type": "string"
                },
                "sisaKamar": {
                    "type": "integer"
                },
                "typeKost": {
                    "type": "string"
                }
            }
        },
        "response.KamarResponse": {
            "type": "object",
            "properties": {
                "hargaKamar": {
                    "type": "integer"
                },
                "nomorKamar": {
                    "type": "string"
                },
                "statusKamar": {
                    "type": "string"
                }
            }
        },
        "response.KostResponse": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "namaKost": {
                    "type": "string"
                },
                "pemilik": {
                    "type": "string"
                },
                "typeKost": {
                    "type": "string"
                }
            }
        },
...
        "response.ViewKostResponse": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "namaKost": {
                    "type": "string"
                },
                "pemilik": {
                    "type": "string"
                },
                "typeKost": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}